refactor: removed global batch id param
